mcproxy: Add support for multiple versions of queries.
Currently always v3 queries are sent. This breaks functionality
of igmp when hosts don't support v3. This is easily reproducible
with macbook as one of the host and a v3 query is sent to the macbook,
causing it to ignore it since it doesn't accept it when its configured
for v2.
More information here:
https://serverfault.com/questions/819793/os-x-network-stack-ignores-igmp-membership-queries